Add RGB to HSV conversion
