+++ src/colorutils.ts
@@ -9,6 +9,7 @@ export { default as hexToRgb } from './hexToRgb';
 export { default as hsvToRgb } from './hsvToRgb';
 export { default as hexToHsv } from './hexToHsv';
 export { default as hsvToHex } from './hsvToHex';
-export { default as rgbToHex } from './rgbToHex';
 export { default as rgbToCmyk } from './rgbToCmyk';
+export { default as rgbToHex } from './rgbToHex';
+export { default as rgbToHsv } from './rgbToHsv';
 export { default as getPerceivedLuminance } from './getPerceivedLuminance';
